Adds hass-aarlo to Custom Components section.
Hass-Aarlo is an asynchronous Arlo Component for Home Assistant.  The component operates in a similar way to the Arlo web site - it opens a single event stream to the Arlo backend and monitors events and state changes for all base stations, cameras and doorbells in a system.   It is actively developed and is superior to the core component in reliability and feature set.